Adjusted ProfileController to Get and Post user information from and to Db. Commented out EditController, no longer needed.
